text-scan-api/requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
python-multipart==0.0.6
pillow==10.0.0
numpy==1.25.2
torch==2.0.1 --index-url https://download.pytorch.org/whl/cpu
torchvision==0.15.2 --index-url https://download.pytorch.org/whl/cpu
python-doctr==0.7.0
